backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
sqlalchemy>=2.0.23
pyodbc>=4.0.39
python-jose>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.4.2
cryptography>=41.0.5
python-multipart>=0.0.6
httpx>=0.25.1
